# Conflicts:
#	src/main/java/model/classes/ProductDAO.java

+++ src/main/java/view/DeleteProductView.java
@@ -1,9 +1,13 @@
 package view;
 
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.Dialog;
 import javafx.scene.control.TextField;
 
 import java.io.IOException;
+import java.util.Optional;
 
 public class DeleteProductView {
     private MainApp mainApp;
@@ -15,7 +19,28 @@ public class DeleteProductView {
     @FXML
     private void deleteProduct(){
         try {
-            this.mainApp.getEngine().productDao().deleteProduct(productBarcode.getText());
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Vahvistus");
+            alert.setHeaderText("Vahvistus");
+            alert.setContentText("Poistetaanko varmasti?");
+
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.get() == ButtonType.OK){
+                boolean res = this.mainApp.getEngine().productDao().deleteProduct(productBarcode.getText());
+                alert.close();
+                if(res == false){
+                    Alert alert2 = new Alert(Alert.AlertType.ERROR);
+                    alert2.setTitle("Virhe");
+                    alert2.setHeaderText("Virhe");
+                    alert2.setContentText("Tuotetta ei löytynyt");
+                    alert2.showAndWait();
+                }
+
+            } else {
+                alert.close();
+            }
+
+
         }catch (Exception e){
             System.out.println("There was an error");
             e.printStackTrace();
